requirements.txt
SQLAlchemy==1.3.4
alembic==1.0.10
inflection==0.3.1
requests==2.22.0
inflection==0.3.1
python-dotenv==0.10.3
psycopg2-binary==2.8.2
